requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
defusedxml==0.7.1
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
Pygments==2.11.2
PyJWT==2.3.0
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
